Cleared Common warnings.

+++ Source/Common/Trace.cpp
@@ -121,7 +121,7 @@ void CTraceLog::TraceMessage(uint32_t module, uint8_t severity, const char * fil
 {
     CGuard Guard(m_CS);
 
-    for (uint32_t i = 0, n = m_Modules.size(); i < n; i++)
+    for (size_t i = 0, n = m_Modules.size(); i < n; i++)
     {
         m_Modules[i]->Write(module, severity, file, line, function, Message);
     }
@@ -183,7 +183,7 @@ m_FlushFile(FlushFile)
     { /* Clamp file size to 5 MB if it exceeds 2047 or falls short of 3. */
         dwMaxFileSize = 5;
     }
-    m_hLogFile.SetMaxFileSize(dwMaxFileSize * MB);
+    m_hLogFile.SetMaxFileSize(static_cast<uint32_t>(dwMaxFileSize * MB));
 
     m_hLogFile.Open(FileName, eMode);
 }
